+++ isitfit/cli/emailMan.py
@@ -3,6 +3,8 @@ import json
 import logging
 logger = logging.getLogger('isitfit')
 
+from ..utils import IsitfitError
+
 class EmailMan:
 
   def __init__(self, dataType, dataVal):
@@ -15,7 +17,7 @@ class EmailMan:
     self.api_man.register()
 
     # submit POST http request
-    response, dt_now = self.api_man.request(
+    response_json, dt_now = self.api_man.request(
       method='post',
       relative_url='./share/email',
       payload_json={
@@ -24,5 +26,33 @@ class EmailMan:
         'share_email': share_email
       }
     )
-    logger.info("Result of share-email:")
-    logger.info(json.dumps(response))
+
+    # check if error
+    if response_json['isitfitapi_status']['code']=='Email verification in progress':
+        raise IsitfitError(response_json['isitfitapi_status']['description'])
+
+    if response_json['isitfitapi_status']['code']=='error':
+        raise IsitfitError(response_json['isitfitapi_status']['description'])
+
+    if response_json['isitfitapi_status']['code']!='ok':
+        response_str = json.dumps(response_json)
+        raise IsitfitError("Unsupported response from server: %s"%response_str)
+
+    # validate schema
+    from schema import SchemaError, Schema, Optional
+    register_schema_2 = Schema({
+        'from': str,
+        Optional(str): object
+      })
+    try:
+        register_schema_2.validate(response_json['isitfitapi_body'])
+    except SchemaError as e:
+        responseBody_str = json.dumps(response_json['isitfitapi_body'])
+        err_msg = "Received response body: %s. Schema error: %s"%(responseBody_str, str(e))
+        raise IsitfitError(err_msg)
+
+    # otherwise proceed
+    emailFrom = response_json['isitfitapi_body']['from']
+    logger.info("Email sent from %s to: %s"%(emailFrom, ", ".join(share_email)))
+    return
+